Changed the saving of the loss_ratios
