+++ openquake/calculators/base.py
@@ -54,6 +54,19 @@ class AssetSiteAssociationError(Exception):
 rlz_dt = numpy.dtype([('uid', (bytes, 200)), ('weight', float)])
 
 
+def set_array(longarray, shortarray):
+    """
+    :param longarray: an array of length L >= l
+    :param shortarray: an array of length l
+
+    Fill `longarray` with the values of `shortarray`, starting from the left.
+    If `shortarry` is shorter than `longarray`, then the remaining elements on
+    the right are filled with `numpy.nan` values.
+    """
+    longarray[:len(shortarray)] = shortarray
+    longarray[len(shortarray):] = numpy.nan
+
+
 class BaseCalculator(with_metaclass(abc.ABCMeta)):
     """
     Abstract base class for all calculators.
@@ -327,14 +340,16 @@ class HazardCalculator(BaseCalculator):
                  for cb in rm.curve_builders if cb.user_provided]
         if not pairs:
             return
+        loss_ratios = numpy.zeros(len(rm), numpy.dtype(pairs))
         for cb in rm.curve_builders:
             if cb.user_provided:
+                loss_ratios_lt = loss_ratios[cb.loss_type]
                 for i, imt_taxo in enumerate(sorted(rm)):
-                    loss_ratios = numpy.array(
-                        rm[imt_taxo].loss_ratios[cb.loss_type])
-                    key = 'loss_ratios/%s-%s-%s' % (imt_taxo + (cb.loss_type,))
-                    self.datastore[key] = loss_ratios
-                    self.datastore[key].attrs['nbytes'] = loss_ratios.nbytes
+                    set_array(loss_ratios_lt[i],
+                              rm[imt_taxo].loss_ratios[cb.loss_type])
+        self.datastore['loss_ratios'] = loss_ratios
+        self.datastore['loss_ratios'].attrs['imt_taxos'] = sorted(rm)
+        self.datastore['loss_ratios'].attrs['nbytes'] = loss_ratios.nbytes
 
     def read_risk_data(self):
         """
