+++ openquake/calculators/tests/classical_risk_test.py
@@ -1,3 +1,4 @@
+import numpy
 from nose.plugins.attrib import attr
 
 from openquake.qa_tests_data.classical_risk import (
@@ -14,8 +15,10 @@ class ClassicalRiskTestCase(CalculatorTestCase):
         out = self.run_calc(case_1.__file__, 'job_risk.ini', exports='xml')
 
         # check loss ratios
-        lrs = self.calc.datastore['loss_ratios/PGA-VF-structural'].value
-        got = scientificformat(lrs, '%.2f')
+        lrs = self.calc.datastore['loss_ratios']
+        self.assertEqual(lrs.dtype.names, ('structural',))
+        numpy.testing.assert_equal(lrs.attrs['imt_taxos'], [['PGA', 'VF']])
+        got = scientificformat(lrs['structural'][0], '%.2f')
         self.assertEqual(got, '0.00 0.01 0.02 0.03 0.04 0.05 0.06 0.07 0.08 '
                          '0.09 0.10 0.12 0.14 0.16 0.18 0.20 0.24 0.28 0.32 '
                          '0.36 0.40 0.48 0.56 0.64 0.72 0.80 0.84 0.88 0.92 '
